requirements.txt
fastapi>=0.129.0
uvicorn>=0.30.0
sqlalchemy>=2.0.30
python-jose==3.3.0
python-multipart>=0.0.9
firebase-admin==6.0.0
redis==5.0.0
celery==5.3.0
boto3==1.26.0
googlemaps==4.10.0
reportlab>=4.0.0
google-genai>=1.0.0
google-cloud-aiplatform>=1.38.0
sentry-sdk>=2.50.0
